updated Last signature to be 64 bit on value param

+++ pyboy/utils.pxd
@@ -63,5 +63,5 @@ cdef class MemoryScanner:
     cdef BCDConverter bcd_converter
 
     cpdef uint64_t read_memory(self, int, int byte_width=*, value_type=*, endian_type=*) noexcept
-    cpdef list scan_memory(self, int, int, int, compare_type=*, value_type=*,int byte_width=*,endian_type=*) noexcept
+    cpdef list scan_memory(self, int, int, uint64_t, compare_type=*, value_type=*,int byte_width=*,endian_type=*) noexcept
     cpdef bint _check_value(self, uint64_t, uint64_t, int)noexcept
